requirements.txt
Django==5.2.8
django-colorfield==0.14.0
gunicorn==23.0.0

# Blog parsing
beautifulsoup4==4.14.3
feedparser==6.0.12

# Documentation pages
mistune==3.1.4
Pygments==2.19.2
python-frontmatter==1.1.0

# REST
djangorestframework==3.16.1
django-filter==25.2
drf-spectacular==0.28.0
drf-orjson-renderer==1.7.3

# Monitoring
django-prometheus==2.4.1

# dev
django-debug-toolbar==5.2.0
django-extensions==4.1
pandas==2.3.3
psutil==7.1.3
psycopg2==2.9.11
rds2py==0.8.0
scipy==1.16.3
h5py==3.15.1

